updating plugins, WP 4.9.5 and parent theme

+++ wp-content/plugins/cornerstone/includes/classes/content/class-content.php
@@ -5,6 +5,7 @@ class Cornerstone_Content {
   protected $id = null;
   protected $title;
   protected $post_type = 'post';
+  protected $post_status = 'post';
   protected $permalink = '';
   protected $data = array();
   // protected $new;
@@ -28,12 +29,11 @@ class Cornerstone_Content {
   protected function create_new( $data ) {
 
     $this->set_title( isset( $data['title'] ) ? $data['title'] : false );
-    // TODO: Populate from incoming data.
+    $this->set_post_type( isset( $data['post_type'] ) ? $data['post_type'] : 'page' );
+    $this->set_post_status( isset( $data['post_status'] ) ? $data['post_status'] : 'draft' );
 
-    $this->set_data( array(
-      'elements' => isset( $data['elements'] ) ? $data['elements'] : array(),
-      'settings' => isset( $data['settings'] ) ? $data['settings'] : array()
-    ) );
+    $this->set_elements( isset( $data['elements'] ) ? $data['elements'] : array( 'data' => '') );
+    $this->set_settings( isset( $data['settings'] ) ? $data['settings'] : array() );
 
   }
 
@@ -44,12 +44,13 @@ class Cornerstone_Content {
     }
 
     if ( ! is_a( $post, 'WP_POST' ) ) {
-      throw new Exception( 'Unable to load header from post.' );
+      throw new Exception( 'Unable to load content from post.' );
     }
 
     $this->id = $post->ID;
     $this->set_title( $post->post_title ? $post->post_title : '' );
     $this->post_type = $post->post_type;
+    $this->post_status = $post->post_status;
 
     $wpml = CS()->loadComponent('Wpml');
     $wpml->before_get_permalink();
@@ -57,7 +58,7 @@ class Cornerstone_Content {
     $wpml->after_get_permalink();
 
 
-    $elements = cs_get_serialized_post_meta( $this->id, '_cornerstone_data', true );
+    $elements = cs_get_serialized_post_meta( $this->id, '_cornerstone_data', true, 'cs_content_load_serialized_content' );
 
     if ( ! is_array( $elements ) ) {
       $elements = array( 'data' => '' );
@@ -70,10 +71,8 @@ class Cornerstone_Content {
       $settings = array();
     }
 
-    $this->set_data( array(
-      'elements' => $this->normalize_elements( $elements ),
-      'settings' => $this->normalize_settings( $settings )
-    ) );
+    $this->set_elements( $elements );
+    $this->set_settings( $settings );
 
   }
 
@@ -95,19 +94,36 @@ class Cornerstone_Content {
     return $settings;
   }
 
-  protected function set_data( $data ) {
-
-    $data = wp_parse_args( is_array( $data ) ? $data : array(), array(
-      'elements' => array(),
-      'settings' => array(),
-    ) );
-
-    $this->set_elements( $data['elements'] );
-    $this->set_settings( $data['settings'] );
-
-  }
-
   public function save() {
+
+    $post_type_object = get_post_type_object( $this->post_type );
+		$caps = (array) $post_type_object->cap;
+
+    $authorized = is_null( $this->id ) ? current_user_can( $caps['create_posts'] ) : current_user_can( $caps['edit_post'], $this->id );
+
+    if ( ! $authorized ) {
+      throw new Exception( 'Unauthorized' );
+    }
+
+    if ( is_null( $this->id) ) {
+
+      $id = wp_insert_post( array(
+        'post_title' => $this->title,
+        'post_type' => $this->post_type,
+        'post_status' => $this->post_status,
+      ));
+
+      if ( is_wp_error( $id ) ) {
+        return $id;
+      }
+
+      if ( 0 === $id ) {
+        return new WP_Error('cs-content', "Unable to save content: $id");
+      }
+
+      $this->id = $id;
+
+    }
 
     $settings = $this->update_settings( $this->get_settings() );
 
@@ -115,10 +131,11 @@ class Cornerstone_Content {
       throw new Exception( 'Error saving content settings: ' . $settings->get_error_message() );
     }
 
-    $elements = $this->update_elements( $this->get_elements() );
+    $elements = $this->get_elements();
+    $update_elements = $this->update_elements( $elements['data'] );
 
-    if ( is_wp_error( $elements ) ) {
-      throw new Exception( 'Error saving content elements: ' . $elements->get_error_message() );
+    if ( is_wp_error( $update_elements ) ) {
+      throw new Exception( 'Error saving content elements: ' . $update_elements->get_error_message() );
     }
 
     return $this->serialize();
@@ -135,7 +152,7 @@ class Cornerstone_Content {
 
   public function get_elements() {
     if ( ! isset( $this->data['elements'] ) ) {
-      $this->data['elements'] = array( 'data' => '');
+      $this->data['elements'] = array( 'data' => '' );
     }
     return $this->data['elements'];
   }
@@ -162,23 +179,43 @@ class Cornerstone_Content {
     return $this->title = sanitize_text_field( $title, sprintf( csi18n('common.untitled-entity'), csi18n('common.entity-content') ) );
   }
 
+  public function set_post_type( $post_type ) {
+    $this->post_type = $post_type;
+  }
+
+  public function set_post_status( $post_status ) {
+    $this->post_status = $post_status;
+  }
+
   public function set_settings( $settings ) {
-    $this->data['settings'] = $settings;
+    $this->data['settings'] = $this->normalize_settings( $settings );
   }
 
   public function set_elements( $elements ) {
-    $this->data['elements'] = $elements;
+    $this->data['elements'] = $this->normalize_elements( $elements );
   }
 
   public function delete() {
+
+    if ( ! current_user_can( CS()->common()->get_post_type_capability( $this->post_type, 'delete_posts' ), $this->id ) ) {
+      throw new Exception( 'Unauthorized' );
+    }
+
     do_action('cornerstone_delete_content', $this->id );
+
     return wp_delete_post( $this->id, true );
+
   }
 
 
   public function update_elements( $elements ) {
 
+    if ( ! $elements ) {
+      return;
+    }
+
     CS()->loadComponent( 'Element_Orchestrator' )->load_elements();
+
     $output = $this->build_output( $elements );
 
 		if ( is_wp_error( $output ) ) {
@@ -190,7 +227,7 @@ class Cornerstone_Content {
 		$old_version = get_post_meta( $this->id, '_cornerstone_version', true );
     update_post_meta( $this->id, '_cornerstone_version', CS()->version() );
 
-		cs_update_serialized_post_meta( $this->id, '_cornerstone_data', $output['data'] );
+		cs_update_serialized_post_meta( $this->id, '_cornerstone_data', $output['data'], '', false, 'cs_content_update_serialized_content' );
 		delete_post_meta( $this->id, '_cornerstone_override' );
     delete_post_meta( $this->id, '_cs_generated_styles');
 
@@ -199,7 +236,9 @@ class Cornerstone_Content {
 		$id = wp_update_post( array(
 			'ID'           => $this->id,
       'title'        => $this->get_title(),
-			'post_content' => wp_slash( $post_content ),
+      'post_type'    => $this->post_type,
+      'post_status'  => $this->post_status,
+      'post_content' => wp_slash( $post_content ),
 		) );
 
     if ( is_wp_error( $id ) ) {
@@ -227,7 +266,7 @@ class Cornerstone_Content {
     $buffer = '';
     $sanitized = array();
 
-    $elements = CS()->loadComponent('Regions')->populate_modules('content', $elements, 'content');
+    $elements = CS()->loadComponent('Regions')->populate_modules( $this->id, $elements, 'content');
 
     foreach ( $elements as $element ) {
       $output = $this->build_element_output( $element );
@@ -373,7 +412,7 @@ class Cornerstone_Content {
 			return null;
 		}
 
-		return $sectionManager->save( $data );
+		return $sectionManager->save( $data, $this );
 
 	}
