+++ wp-content/plugins/cornerstone/includes/classes/footers/class-footer.php
@@ -17,7 +17,21 @@ class Cornerstone_Footer {
         $this->create_new( $post );
       }
     } else {
-      $this->load_from_post( $post );
+
+      if ( is_a( $post, 'Cornerstone_Template' ) ) {
+
+        $meta = $post->get_meta();
+
+        $this->create_new( array(
+          'title'   => $post->get_title(),
+          'regions' => isset( $meta['regions'] ) ? $meta['regions'] : array(),
+          'settings' => isset( $meta['settings'] ) ? $meta['settings'] : array()
+        ) );
+
+      } else {
+        $this->load_from_post( $post );
+      }
+
     }
 
   }
@@ -48,9 +62,11 @@ class Cornerstone_Footer {
       throw new Exception( 'Attempted to load footer from incorrect post_type.' );
     }
 
+
     $this->id = $post->ID;
     $this->set_title( $post->post_title ? $post->post_title : '' );
-    $this->set_content( cs_maybe_json_decode( $post->post_content ) );
+    $content = apply_filters('cs_footer_load_content', $post->post_content );
+    $this->set_content( cs_maybe_json_decode( $content ) );
 
   }
 
@@ -68,14 +84,29 @@ class Cornerstone_Footer {
 
   public function save() {
 
+    if ( ! current_user_can( 'manage_options' ) ) {
+      throw new Exception( 'Unauthorized' );
+    }
+
+    // Expand Template
+    $settings = $this->get_settings();
+
+    if ( isset( $settings['_template'] ) && $settings['_template'] ) {
+      $template = new Cornerstone_Template( $settings['_template'] );
+      $meta = $template->get_meta();
+      $this->set_regions( isset( $meta['regions'] ) ? $meta['regions'] : array() );
+      $this->set_settings( isset( $meta['settings'] ) ? $meta['settings'] : array() );
+    }
+
+
     $args = array(
       'post_title'   => $this->get_title(),
       'post_type'    => 'cs_footer',
       'post_status'  => 'tco-data',
-      'post_content' => wp_slash( cs_json_encode( array(
+      'post_content' => apply_filters('cs_footer_update_content', wp_slash( cs_json_encode( array(
         'regions' => $this->get_regions(),
         'settings' => $this->get_settings()
-      ) ) )
+      ) ) ) )
     );
 
     if ( is_int( $this->id ) ) {
@@ -154,7 +185,14 @@ class Cornerstone_Footer {
   }
 
   public function delete() {
+
+    if ( ! current_user_can( 'manage_options' ) ) {
+      throw new Exception( 'Unauthorized' );
+    }
+
     do_action('cornerstone_delete_footer', $this->id );
+
     return wp_delete_post( $this->id, true );
+
   }
 }
